v1 bis

+++ CoffeeMachine.h
@@ -13,7 +13,7 @@ private:
 
 public:
     CoffeeMachine();
-    void insererPiece(int piece); //uniquement piece 0,5
+    void insererPiece(float piece); //uniquement piece 0,5
     void rechargerEau(); //recharge la capacité d'eau
     void rechargerCafe(); //recharge la capacité de café
     void servir(); //sert un café de 0,15cl d'eau et de 10g de café coût du café est 0.5 euros
